# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Video Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type VideoStatus = typeof VideoStatus[keyof typeof VideoStatus];


export const VideoStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type VideoAspectRatio = typeof VideoAspectRatio[keyof typeof VideoAspectRatio];


export const VideoAspectRatio = {
  '16:9': '16:9',
  '9:16': '9:16',
} as const;

export interface Video {
  id: number;
  prompt: string;
  status: VideoStatus;
  aspectRatio: VideoAspectRatio;
  /** Duration in seconds (4, 6, or 8) */
  duration: number;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  falRequestId?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type VideoInputAspectRatio = typeof VideoInputAspectRatio[keyof typeof VideoInputAspectRatio];


export const VideoInputAspectRatio = {
  '16:9': '16:9',
  '9:16': '9:16',
} as const;

export type VideoInputDuration = typeof VideoInputDuration[keyof typeof VideoInputDuration];


export const VideoInputDuration = {
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_8: 8,
} as const;

export interface VideoInput {
  /** @minLength 3 */
  prompt: string;
  aspectRatio?: VideoInputAspectRatio;
  duration?: VideoInputDuration;
}

export interface UnlockInput {
  code: string;
}

export type PlanInfoPlan = typeof PlanInfoPlan[keyof typeof PlanInfoPlan];


export const PlanInfoPlan = {
  free: 'free',
  premium: 'premium',
  admin: 'admin',
} as const;

export interface PlanInfo {
  plan: PlanInfoPlan;
}

export type QuotaInfoPlan = typeof QuotaInfoPlan[keyof typeof QuotaInfoPlan];


export const QuotaInfoPlan = {
  free: 'free',
  premium: 'premium',
  admin: 'admin',
} as const;

export interface QuotaInfo {
  plan: QuotaInfoPlan;
  used: number;
  /** @nullable */
  limit: number | null;
  canGenerate: boolean;
  allowedDurations: number[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export type OpenaiImageInputSize = typeof OpenaiImageInputSize[keyof typeof OpenaiImageInputSize];


export const OpenaiImageInputSize = {
  '1024x1024': '1024x1024',
  '1536x1024': '1536x1024',
  '1024x1536': '1024x1536',
} as const;

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}

export interface OpenaiImageOutput {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export interface VideoStats {
  total: number;
  pending: number;
  processing: number;
  completed: number;
  failed: number;
}