# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Video Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: videos
    description: Video generation operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all generated videos
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: generateVideo
      tags: [videos]
      summary: Submit a video generation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Video generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Video generation service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /videos/{id}/status:
    get:
      operationId: getVideoStatus
      tags: [videos]
      summary: Poll video generation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiImageOutput"

  /unlock:
    post:
      operationId: unlockPlan
      tags: [videos]
      summary: Unlock a plan using an access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockInput"
      responses:
        "200":
          description: Plan unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanInfo"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /quota:
    get:
      operationId: getQuota
      tags: [videos]
      summary: Get current plan quota and usage
      parameters:
        - name: x-plan-code
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Quota info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaInfo"

  /videos/stats:
    get:
      operationId: getVideoStats
      tags: [videos]
      summary: Get dashboard stats (total, pending, completed, failed)
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Video:
      type: object
      required: [id, prompt, status, aspectRatio, duration, createdAt]
      properties:
        id:
          type: integer
        prompt:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        aspectRatio:
          type: string
          enum: ["16:9", "9:16"]
        duration:
          type: integer
          description: Duration in seconds (4, 6, or 8)
        videoUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        falRequestId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]

    VideoInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 3
        aspectRatio:
          type: string
          enum: ["16:9", "9:16"]
          default: "16:9"
        duration:
          type: integer
          enum: [2, 3, 4, 5, 8]
          default: 4

    UnlockInput:
      type: object
      required: [code]
      properties:
        code:
          type: string

    PlanInfo:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [free, premium, admin]

    QuotaInfo:
      type: object
      required: [plan, used, limit, canGenerate, allowedDurations]
      properties:
        plan:
          type: string
          enum: [free, premium, admin]
        used:
          type: integer
        limit:
          type: ["integer", "null"]
        canGenerate:
          type: boolean
        allowedDurations:
          type: array
          items:
            type: integer

    OpenaiConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
    OpenaiConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
    OpenaiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
    OpenaiConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    OpenaiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    OpenaiImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
    OpenaiImageOutput:
      type: object
      required: [b64_json]
      properties:
        b64_json:
          type: string
    OpenaiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    VideoStats:
      type: object
      required: [total, pending, processing, completed, failed]
      properties:
        total:
          type: integer
        pending:
          type: integer
        processing:
          type: integer
        completed:
          type: integer
        failed:
          type: integer