# lib/api-zod/src/generated/types/openaiImageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Video Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiImageInputSize } from './openaiImageInputSize';

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}